Se corrige ciclo entre servicios, y se cambia manera de agregar usuarios
