+++ src/main/java/com/example/SAC/controller/RegistroController.java
@@ -5,7 +5,6 @@ import com.example.SAC.entity.Propietario;
 import com.example.SAC.entity.Residente;
 import com.example.SAC.service.AdministradorService;
 import com.example.SAC.service.PropietarioService;
-import com.example.SAC.service.RegistroService;
 import com.example.SAC.service.ResidenteService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
@@ -14,23 +13,29 @@ import org.springframework.web.bind.annotation.*;
 @RequestMapping("/api/register")
 public class RegistroController {
     @Autowired
-    RegistroService registroService;
+    ResidenteService residenteService;
+
+    @Autowired
+    PropietarioService propietarioService;
+
+    @Autowired
+    AdministradorService administradorService;
 
     //Registrar residente
     @PostMapping("/residente")
     public Residente registrarResidente(@RequestBody Residente residente) {
-        return registroService.registrarResidente(residente);
+        return residenteService.registrarResidente(residente);
     }
     //Registrar propietario
     @PostMapping("/propietario")
     public Propietario registrarPropietario(@RequestBody Propietario propietario) {
-        return registroService.registrarPropietario(propietario);
+        return propietarioService.registrarPropietario(propietario);
     }
 
     //Registrar admin
     @PostMapping("/1603aeZakmi/administrador")
     public Administrador registrarAdministrador(@RequestBody Administrador administrador) {
-        return registroService.registrarAdministrador(administrador);
+        return administradorService.registrarAdministrador(administrador);
     }
