+++ src/main/java/com/example/SAC/service/AdministradorService.java
@@ -1,14 +1,15 @@
 package com.example.SAC.service;
 
-import com.example.SAC.dto.UsuarioDTO;
 import com.example.SAC.entity.Administrador;
 import com.example.SAC.entity.Cuenta;
-import com.example.SAC.entity.Residente;
 import com.example.SAC.repository.AdministradorRepository;
 import com.example.SAC.repository.CuentaRepository;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+import org.springframework.web.server.ResponseStatusException;
 
 import java.util.Optional;
 import java.util.List;
@@ -27,11 +28,32 @@ public class AdministradorService {
     @Autowired
     private PropietarioService propietarioService;
     @Autowired
-    private RegistroService registroService;
+    private ValidacionService validacionService;
+
+    private String normalize(String s) {
+        if (s == null) return null;
+        String t = s.trim();
+        return t.isEmpty() ? null : t.toLowerCase();
+    }
 
 
+    //Agregar administrador -- Registrar administrador =D
+    @Transactional
+    public Administrador registrarAdministrador(Administrador administrador) {
+        String correo = normalize(administrador.getCorreo());
+        String documento = normalize(administrador.getDocumento());
+        String telefono = normalize(administrador.getTelefono());
 
-    public Administrador agregarAdministrador(Administrador administrador){
+        if (correo != null && validacionService.verificarCorreoRegistrado(correo)) {
+            throw new ResponseStatusException(HttpStatus.CONFLICT, "El correo ya se encuentra registrado");
+        }
+        if (documento != null && validacionService.verificarDocumentoRegistrado(documento)) {
+            throw new ResponseStatusException(HttpStatus.CONFLICT, "El documento ya se encuentra registrado");
+        }
+        if (telefono != null && validacionService.verificarNumeroRegistrado(telefono)) {
+            throw new ResponseStatusException(HttpStatus.CONFLICT, "El número de teléfono ya se encuentra registrado");
+        }
+
         //Se crea la cuenta
         Cuenta cuentanueva = new Cuenta();
         //Se define el tipo de cuenta que se crea
@@ -55,19 +77,19 @@ public class AdministradorService {
         try {
             // Validar correo solo si cambia y ya está registrado por otro
             if (!administrador.getCorreo().equals(nuevosDatos.getCorreo())
-                    && registroService.verificarCorreoRegistrado(nuevosDatos.getCorreo())) {
+                    && validacionService.verificarCorreoRegistrado(nuevosDatos.getCorreo())) {
                 throw new RuntimeException("El correo ya está registrado");
             }
 
             // Validar documento solo si cambia y ya está registrado por otro
             if (!administrador.getDocumento().equals(nuevosDatos.getDocumento())
-                    && registroService.verificarDocumentoRegistrado(nuevosDatos.getDocumento())) {
+                    && validacionService.verificarDocumentoRegistrado(nuevosDatos.getDocumento())) {
                 throw new RuntimeException("El documento ya está registrado");
             }
 
             // Validar teléfono solo si cambia y ya está registrado por otro
             if (!administrador.getTelefono().equals(nuevosDatos.getTelefono())
-                    && registroService.verificarNumeroRegistrado(nuevosDatos.getTelefono())) {
+                    && validacionService.verificarNumeroRegistrado(nuevosDatos.getTelefono())) {
                 throw new RuntimeException("El número ya está registrado");
             }
